goalFinder/pyproject.toml
[project]
name = "goalfinder"
version = "0.1.0"
description = "Goal extraction MVP scaffold for a FastAPI and React app"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "google-genai>=1.66.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
    "websockets>=16.0",
]
